Added home-manager profiles (#30)
* feat: added home-manager profiles for non-nixos setups

* feat(lib): add homeDirectory variable for Darwin systems

Introduce a new variable `homeDirectory` to handle user home directory paths. This variable checks if the system is either "aarch64-darwin" or "x86_64-darwin" and sets the home directory to "/Users/${username}". For other systems, it defaults to "${defaultHomePath}/${username}". Update `homeManagerShared` to include `homeDirectory`.

* fix: correct path for k9s skin configuration file

The source path for the k9s skin.yml file was updated to point to the correct relative location.

* refactor(lib): simplify conditional imports for home-manager

Simplified the conditional logic for importing home-manager and home-manager-desktop configurations. Combined conditions to reduce redundancy and improve readability.

* chore: remove vscode program enablement from home-manager-desktop.nix

The configuration for enabling the vscode program has been removed from the home-manager-desktop.nix file. This change may affect users who rely on vscode being enabled by default in this configuration.